Add 'Accept-Encoding' to CORS header whitelist
Safari sends the Accept-Encoding header in its preflight HTTP request to
the server. Allowing the header enables the server to accept the
request. Specifically, this change allows POST /api/cart to work
properly: the server sends an appropriate response code, allowing the
frontend to get the updated cart contents successfully.
[Finishes #119862935]