WIP: Partial implementation of /orders API
This change defines the mongoose model for orders and partially
implements the first few API routes, most notably POST /orders. So
far, calling that API will create a new order if the paymentMethod is
cash and all the fields the user entered are valid.

Paypal implementation is not done yet as of this change. [#119419923]